fix(enterprise mgmt): fix integration tests
